+++ at_persistence/at_persistence_spec/lib/at_persistence_spec.dart
@@ -1,10 +1,10 @@
 library at_persistence_spec;
 
+export 'package:at_persistence_spec/src/compaction/at_compaction_strategy.dart';
 export 'package:at_persistence_spec/src/exception/exceptions.dart';
 export 'package:at_persistence_spec/src/keystore/keystore.dart';
 export 'package:at_persistence_spec/src/keystore/keystore_manager.dart';
-export 'package:at_persistence_spec/src/keystore/secondary_keystore.dart';
-export 'package:at_persistence_spec/src/compaction/at_compaction_strategy.dart';
 export 'package:at_persistence_spec/src/keystore/log_keystore.dart';
-export 'package:at_persistence_spec/src/log/at_commit_log_manager.dart';
+export 'package:at_persistence_spec/src/keystore/secondary_keystore.dart';
 export 'package:at_persistence_spec/src/log/at_access_log_manager.dart';
+export 'package:at_persistence_spec/src/log/at_commit_log_manager.dart';