+++ at_secondary/at_persistence_secondary_server/lib/src/keystore/hive_keystore.dart
@@ -198,9 +198,9 @@ class HiveKeystore implements SecondaryKeyStore<String, AtData?, AtMetaData?> {
     try {
       var expiredKeys = await getExpiredKeys();
       if (expiredKeys.isNotEmpty) {
-        expiredKeys.forEach((element) {
-          remove(element);
-        });
+        for (var element in expiredKeys) {
+          await remove(element);
+        }
         result = true;
       }
     } on Exception catch (e) {
@@ -361,8 +361,9 @@ class HiveKeystore implements SecondaryKeyStore<String, AtData?, AtMetaData?> {
   /// Returns true if key exists in [HiveKeystore]; else false.
   @override
   bool isKeyExists(String key) {
-    var result = persistenceManager.getBox().containsKey(key);
-    return result;
+    return persistenceManager
+        .getBox()
+        .containsKey(keyStoreHelper.prepareKey(key));
   }
 
   ///Restarts the hive box.