+++ at_secondary/at_secondary_server/bin/main.dart
@@ -10,10 +10,8 @@ import 'package:at_utils/at_logger.dart';
 /// Throws [ArgParserException] for invalid arguments passed.
 /// @ param - List<String> atSign and port
 Future<void> main(List<String> arguments) async {
-  var isDebug = AtSecondaryConfig.debugLog!;
-  if (isDebug) {
-    AtSignLogger.root_level = 'finest';
-  }
+  AtSignLogger.root_level = AtSecondaryConfig.logLevel;
+
   var logger = AtSignLogger('main');
 
   try {