+++ at_secondary/at_secondary_server/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.0.11
+- Changes to support reset of ttb and ttl
+## 3.0.10
+- Workaround for signing private key not found issue.
+## 3.0.9
+- Enhance commit log compaction service.
+- Notification expiry feature
 ## 3.0.8
 - reduce compaction interval to 12 hrs
 - compaction delete bug fix